better organize agents
