+++ lib/steering.dart
@@ -1,3 +1,8 @@
+// ignore_for_file: directives_ordering
+
 // Abstract interfaces
 export 'src/steering/kinematics.dart' show Kinematics;
 export 'src/steering/steerable.dart' show Steerable;
+
+// Kinematics implementations
+export 'src/steering/kinematics/basic_kinematics.dart';
